fix feature/create-diary and feature/update-diary

+++ src/main/java/com/ttc/diary/service/impl/DiaryServiceImpl.java
@@ -1,11 +1,11 @@
 package com.ttc.diary.service.impl;
 
+import com.ttc.diary.exception.ResourceNotFoundException;
+import com.ttc.diary.model.*;
 import com.ttc.diary.model.entity.Diary;
 import com.ttc.diary.model.entity.DiaryImage;
 import com.ttc.diary.model.entity.Topic;
 import com.ttc.diary.model.entity.User;
-import com.ttc.diary.model.*;
-import com.ttc.diary.exception.ResourceNotFoundException;
 import com.ttc.diary.repository.DiaryImageRepository;
 import com.ttc.diary.repository.DiaryRepository;
 import com.ttc.diary.repository.TopicRepository;
@@ -13,9 +13,9 @@ import com.ttc.diary.service.DiaryService;
 import com.ttc.diary.util.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -45,7 +45,8 @@ public class DiaryServiceImpl implements DiaryService {
 
     public DiaryDto createDiary(DiaryDto dto) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (!authentication.isAuthenticated()) throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user");
+        if (!authentication.isAuthenticated())
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user");
         UserPrincipal principal = (UserPrincipal) authentication.getPrincipal();
         Diary diary = new Diary();
         diary.setOwner(new User(principal.getId()));
@@ -57,11 +58,13 @@ public class DiaryServiceImpl implements DiaryService {
 
         diaryRepository.saveAndFlush(diary);
         dto.setId(diary.getId());
-        List<DiaryImage> images = dto.getImageDtos().stream()
-                .map(s-> new DiaryImage(s.getPath(), diary)).collect(Collectors.toList());
-        diaryImageRepository.saveAll(images);
-        for (int i=0; i<images.size(); i++){
-            dto.getImageDtos().get(i).setId(images.get(i).getId());
+        if (dto.getImageDtos() != null && dto.getImageDtos().size() > 0) {
+            List<DiaryImage> images = dto.getImageDtos().stream()
+                    .map(s -> new DiaryImage(s.getPath(), diary)).collect(Collectors.toList());
+            diaryImageRepository.saveAll(images);
+            for (int i = 0; i < images.size(); i++) {
+                dto.getImageDtos().get(i).setId(images.get(i).getId());
+            }
         }
         return dto;
     }
@@ -69,7 +72,7 @@ public class DiaryServiceImpl implements DiaryService {
     @Override
     public ResponseEntity<Diary> changeFavoriteStatus(Long id) {
         Diary diary = diaryRepository.findById(id)
-                .orElseThrow(() -> new IllegalStateException("Can't found diary with id "+id+" !!!"));
+                .orElseThrow(() -> new IllegalStateException("Can't found diary with id " + id + " !!!"));
         diary.setFavorite(!diary.getFavorite());
 
         return null;
@@ -78,7 +81,7 @@ public class DiaryServiceImpl implements DiaryService {
     @Override
     public DiaryDetailDto getDiaryById(Long id) {
         Diary diary = diaryRepository.findById(id)
-                .orElseThrow(() -> new IllegalStateException("Can't found diary with id "+id+" !!!"));
+                .orElseThrow(() -> new IllegalStateException("Can't found diary with id " + id + " !!!"));
 
         DiaryDetailDto diaryDetailDto = new DiaryDetailDto();
 
@@ -113,6 +116,7 @@ public class DiaryServiceImpl implements DiaryService {
         diaryRepository.delete(diary);
         return "Delete success";
     }
+
     @Override
     public DiaryDto updateDiary(Long id, DiaryDto diaryDto) {
         Diary diary = diaryRepository.findById(id)
@@ -122,19 +126,24 @@ public class DiaryServiceImpl implements DiaryService {
 
         diary.setTopics(topicRepository.findAllById(diaryDto.getTopicIds()));
 
-        List<Long> imageIds = diaryDto.getImageDtos().stream().map(ImageDto::getId).collect(Collectors.toList());
-        List<DiaryImage> diaryImages = diaryImageRepository.findAllByDiaryId(id);
+        if (diaryDto.getImageDtos() != null && diaryDto.getImageDtos().size() > 0) {
+            List<Long> imageIds = diaryDto.getImageDtos().stream().map(ImageDto::getId).collect(Collectors.toList());
+            List<DiaryImage> diaryImages = diaryImageRepository.findAllByDiaryId(id);
 
-        List<DiaryImage> diaryOldImages = diaryImages.stream()
-                .filter(s -> !imageIds.contains(s.getId()))
-                .collect(Collectors.toList());
-        diaryImageRepository.deleteInBatch(diaryOldImages);
+            List<DiaryImage> diaryOldImages = diaryImages.stream()
+                    .filter(s -> !imageIds.contains(s.getId()))
+                    .collect(Collectors.toList());
+            diaryImageRepository.deleteInBatch(diaryOldImages);
 
-        List<DiaryImage> diaryNewImages = diaryDto.getImageDtos().stream()
-                .filter(s -> s.getId() == null || (s.getId() != null && s.getId() == 0))
-                .map(s -> new DiaryImage(s.getPath(), diary))
-                .collect(Collectors.toList());
-        diaryImageRepository.saveAll(diaryNewImages);
+            List<DiaryImage> diaryNewImages = diaryDto.getImageDtos().stream()
+                    .filter(s -> s.getId() == null || (s.getId() != null && s.getId() == 0))
+                    .map(s -> new DiaryImage(s.getPath(), diary))
+                    .collect(Collectors.toList());
+            diaryImageRepository.saveAll(diaryNewImages);
+            for (int i = 0; i < diaryNewImages.size(); i++) {
+                diaryDto.getImageDtos().get(i).setId(diaryNewImages.get(i).getId());
+            }
+        }
 
         return diaryDto;
     }
